requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.5
djangorestframework==3.16.1
gunicorn==23.0.0
numpy==2.3.2
packaging==25.0
pandas==2.3.2
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.9.0
